+++ rdxgraphics/GUI/GUI.h
@@ -7,7 +7,7 @@ class GUI : public BaseSingleton<GUI>
 public:
 	static bool Init();
 	static void Terminate();
-	static void Update(double dt);
+	static void Update(float dt);
 	static void Draw();
 
 	inline static entt::entity GetSelectedEntity() { return g.m_SelectedEntity; }
